// src/stores/theme/types/datePickerConfig.ts
export interface DatePickerConfig {
  background: string;
  borderDefault: string;
  borderHover: string;
  borderFocus: string;
  borderError: string;
  textColor: string;
  placeholderColor: string;
  iconColor: string;
  calendarBackground: string;
  calendarHeaderBackground: string;
  calendarHeaderTextColor: string;
  calendarCellHoverBackground: string;
  calendarSelectedBackground: string;
  calendarSelectedTextColor: string;
  calendarTodayBorderColor: string;
  calendarOtherMonthTextColor: string;
  borderRadius: string;
  labelColor: string;
  calendarBorderColor: string;
  calendarShadow: string;
  weekHeaderBorderColor: string;
  weekDayLabelColor: string;
  dayTextColor: string;
  navArrowColor: string;
  selectedBorderColor: string;
  rangeBackground: string;
  rangeTextColor: string;
  rangeSeparatorBackground: string;
  rangeSeparatorBorderColor: string;
  rangeSeparatorIconColor: string;
  focusRingColor: string;
  transitionDuration: string;
  // Typography
  calendarFontFamily: string;
  headerFontSize: string;
  headerFontWeight: string;
  dayFontSize: string;
  dayFontWeight: string;
  weekDayFontSize: string;
  weekDayFontWeight: string;
  inputFontSize: string;
  labelFontSize: string;
  // Sizing & Spacing
  panelWidth: string;
  panelPadding: string;
  panelBorderRadius: string;
  dayCellSize: string;
  headerHeight: string;
  headerPadding: string;
  navButtonSize: string;
  inputHeight: string;
}